fix: case when user switches from any lang(except of en) to another lang(except of en)
